refactor: enforce MAIN dependency direction
